+++ main.rb
@@ -1,4 +1,4 @@
-require_relative 'app'
+require_relative './src/app'
 
 def main
   system('cls') # for windows
